{"schema":"decodex/commit/1","summary":"Split MCP planning tools","authority":"XY-1115"}
